move broker failure-mode tests to integration and tighten assertions
